Add examples for * protected access * the final keyword * and Polymorphism
